refactor(test): audit structural.rs — rename, parameterize, and reorganize tests
Audit structural.rs for naming, comments, coverage, and test placement:

- Rename tests to match error variants: test_boundary_proof_unverifiable
  → test_unexpected_start_proof, test_wrong_end_root_boundary_check
  → test_boundary_proof_rejects_wrong_root
- Add 'crafted' prefix to all tests that construct FrozenChangeProof by hand
- Parameterize test_keys_not_sorted → test_crafted_keys with #[test_case]
  covering both duplicate and reverse-ordered keys
- Tighten assertions: InvalidRange checks exact keys, wrong-root check
  asserts UnexpectedHash instead of broad ProofError wildcard
- Add precondition assert on generator end_proof in test_crafted_missing_end_proof
- Drop duplicate test_delete_range_rejected_with_boundary_proofs (same
  code path as test_crafted_delete_range_rejected) and simplify the
  remaining test
- Update module doc with variant coverage map and cross-references
- Add short comments where test mechanism isn't obvious from the name

Move misplaced Phase 1 structural tests to their thematic homes:
- DeleteRangeFoundInChangeProof (2 tests): edge_cases.rs → structural.rs
- MissingEndProof empty batch_ops: pipeline.rs → structural.rs
- EndProofOperationMismatch (2 tests): pipeline.rs → attack.rs
