+++ firewood/src/merkle/tests/change/structural.rs
@@ -1,9 +1,18 @@
 // Copyright (C) 2025, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE.md for licensing terms.
 
-//! Structural validation: one test per error variant in `verify_change_proof_structure`.
+//! Structural validation of change proofs (Phase 1).
+//!
+//! One test per error variant in [`verify_change_proof_structure`], covering
+//! argument validation, boundary proof presence, key ordering, and hash chain
+//! verification against `end_root`.
+//!
+//! Variants tested elsewhere:
+//! - `StartProofOperationMismatch` — attack.rs (adversarial boundary injection)
+//! - `EndProofOperationMismatch` — attack.rs (adversarial boundary injection)
 
 use super::*;
+use test_case::test_case;
 
 #[test]
 fn test_inverted_range_rejected() {
@@ -16,11 +25,17 @@ fn test_inverted_range_rejected() {
 
     let err = verify_change_proof_structure(&proof, root2, Some(b"\xa0"), Some(b"\x10"), None)
         .unwrap_err();
-    assert!(matches!(err, api::Error::InvalidRange { .. }));
+    assert!(matches!(
+        err,
+        api::Error::InvalidRange { start_key, end_key }
+            if *start_key == *b"\xa0" && *end_key == *b"\x10"
+    ));
 }
 
+/// Start proof present but `requested_start_key` is None — the proof
+/// is unverifiable so the structural check rejects it.
 #[test]
-fn test_boundary_proof_unverifiable() {
+fn test_unexpected_start_proof() {
     let (db, _dir) = setup_db![(b"\x00", b"v0"), (b"\x10", b"v1")];
     let (root1, root2) = setup_2nd_commit!(db, [(b"\x05", b"v2")]);
 
@@ -28,7 +43,6 @@ fn test_boundary_proof_unverifiable() {
         .change_proof(root1, root2.clone(), Some(b"\x00"), Some(b"\x10"), None)
         .unwrap();
 
-    // Non-empty start_proof but start_key=None
     let err = verify_change_proof_structure(&proof, root2, None, Some(b"\x10"), None).unwrap_err();
     assert!(matches!(
         err,
@@ -36,8 +50,10 @@ fn test_boundary_proof_unverifiable() {
     ));
 }
 
-#[test]
-fn test_keys_not_sorted() {
+/// The `key <= prev.key()` check rejects both duplicate and reverse-ordered keys.
+#[test_case(b"\x50", b"\x50" ; "duplicate keys rejected")]
+#[test_case(b"\xa0", b"\x50" ; "reverse order rejected")]
+fn test_crafted_keys(key_a: &[u8], key_b: &[u8]) {
     let (db, _dir) = setup_db![(b"\x10", b"v0"), (b"\xa0", b"v1")];
     let (root1, root2) = setup_2nd_commit!(db, [(b"\x50", b"mid")]);
 
@@ -50,11 +66,11 @@ fn test_keys_not_sorted() {
         crate::Proof::new(valid.end_proof().as_ref().into()),
         Box::new([
             BatchOp::Put {
-                key: b"\x50".to_vec().into(),
+                key: key_a.to_vec().into(),
                 value: b"a".to_vec().into(),
             },
             BatchOp::Put {
-                key: b"\x50".to_vec().into(),
+                key: key_b.to_vec().into(),
                 value: b"b".to_vec().into(),
             },
         ]),
@@ -97,6 +113,7 @@ fn test_end_key_less_than_last_key() {
     ));
 }
 
+/// Proof has 2 batch ops but `max_length` is 1.
 #[test]
 fn test_proof_larger_than_max_length() {
     let (db, _dir) = setup_db![(b"\x10", b"v0"), (b"\xa0", b"v1")];
@@ -113,8 +130,10 @@ fn test_proof_larger_than_max_length() {
     ));
 }
 
+/// Non-empty `batch_ops` with `requested_start_key` and `requested_end_key`
+/// but both boundary proofs empty.
 #[test]
-fn test_missing_boundary_proof() {
+fn test_crafted_missing_boundary_proof() {
     let proof = FrozenChangeProof::new(
         crate::Proof::new(Box::new([])),
         crate::Proof::new(Box::new([])),
@@ -137,14 +156,19 @@ fn test_missing_boundary_proof() {
     ));
 }
 
+/// Non-empty `batch_ops` with start proof present but end proof stripped.
 #[test]
-fn test_missing_end_proof() {
+fn test_crafted_missing_end_proof() {
     let (db, _dir) = setup_db![(b"\x10", b"v0"), (b"\xa0", b"v1")];
     let (root1, root2) = setup_2nd_commit!(db, [(b"\x50", b"mid")]);
 
     let valid = db
         .change_proof(root1, root2.clone(), None, None, None)
         .unwrap();
+    assert!(
+        !valid.end_proof().is_empty(),
+        "generator must produce end proof for non-empty ops"
+    );
 
     let crafted = FrozenChangeProof::new(
         crate::Proof::new(valid.start_proof().as_ref().into()),
@@ -158,18 +182,17 @@ fn test_missing_end_proof() {
     ));
 }
 
+/// Non-empty end proof with empty `batch_ops` and no `requested_end_key`.
 #[test]
-fn test_unexpected_end_proof() {
+fn test_crafted_unexpected_end_proof() {
     let (db, _dir) = setup_db![(b"\x10", b"v0"), (b"\x20", b"v1")];
     let (root1, root2) = setup_2nd_commit!(db, [(b"\x20", b"changed")]);
 
-    // Get a proof with a non-empty end_proof
     let valid = db
         .change_proof(root1, root2.clone(), None, Some(b"\x20"), None)
         .unwrap();
     assert!(!valid.end_proof().is_empty());
 
-    // Craft: non-empty end_proof + empty batch_ops + end_key=None
     let crafted = FrozenChangeProof::new(
         crate::Proof::new(Box::new([])),
         crate::Proof::new(valid.end_proof().as_ref().into()),
@@ -183,8 +206,13 @@ fn test_unexpected_end_proof() {
     ));
 }
 
+/// Boundary proof hash chain was built for `root2` but verified against
+/// `HashKey::empty()` — the root hash in the proof nodes won't match,
+/// producing `UnexpectedHash`.
+///
+/// TODO: `ProofError::UnexpectedHash` is poorly named. The hash isn't unexpected, it's just wrong.
 #[test]
-fn test_wrong_end_root_boundary_check() {
+fn test_boundary_proof_rejects_wrong_root() {
     let (db, _dir) = setup_db![(b"\x10", b"v0"), (b"\xa0", b"v1")];
     let (root1, root2) = setup_2nd_commit!(db, [(b"\x50", b"mid")]);
 
@@ -199,5 +227,48 @@ fn test_wrong_end_root_boundary_check() {
         None,
     )
     .unwrap_err();
-    assert!(matches!(err, api::Error::ProofError(_)));
+    assert!(matches!(
+        err,
+        api::Error::ProofError(crate::ProofError::UnexpectedHash)
+    ));
+}
+
+#[test]
+fn test_crafted_delete_range_rejected() {
+    let (db, _dir) = setup_db![(b"\x10", b"v0"), (b"\xa0", b"v1")];
+    let (root1, root2) = setup_2nd_commit!(db, [(b"\x50", b"mid")]);
+
+    let valid = db
+        .change_proof(root1, root2.clone(), None, None, None)
+        .unwrap();
+
+    let crafted = FrozenChangeProof::new(
+        crate::Proof::new(valid.start_proof().as_ref().into()),
+        crate::Proof::new(valid.end_proof().as_ref().into()),
+        Box::new([BatchOp::DeleteRange {
+            prefix: b"\x50".to_vec().into(),
+        }]),
+    );
+
+    let err = verify_change_proof_structure(&crafted, root2, None, None, None).unwrap_err();
+    assert!(matches!(
+        err,
+        api::Error::ProofError(crate::ProofError::DeleteRangeFoundInChangeProof)
+    ));
+}
+
+#[test]
+fn test_crafted_missing_end_proof_empty_batch_ops() {
+    let proof = FrozenChangeProof::new(
+        crate::Proof::new(Box::new([])),
+        crate::Proof::new(Box::new([])),
+        Box::new([]),
+    );
+    let err =
+        verify_change_proof_structure(&proof, api::HashKey::empty(), None, Some(b"\x50"), None)
+            .unwrap_err();
+    assert!(matches!(
+        err,
+        api::Error::ProofError(crate::ProofError::MissingEndProof)
+    ));
 }
